Salary Prediction/requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
category_encoders
scikit-learn
imbalanced-learn
xgboost
statsmodels
pymc3
arviz
